% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBubbleCluster.R
\name{add_Bubble_Cluster}
\alias{add_Bubble_Cluster}
\title{addBubbleCluster}
\usage{
add_Bubble_Cluster(
  m,
  data,
  lng,
  lat,
  method = "mean",
  label,
  FiltColName,
  FiltColVal,
  color = "#FF0000",
  InnerCircle = list(use = T, radius = 12000, stroke, alpha = 0.5, color = "#FF0000"),
  OuterCircle = list(use = T, radius = 16000, stroke, alpha = 0.4, color = "#FF0000")
)
}
\arguments{
\item{m}{leaflet map object.}

\item{data}{your raw data to use. it should be a data frame}

\item{lng}{column name for longitude. WGS84 coordinate.}

\item{lat}{column name for latitude.  WGS84 coordinate.}

\item{method}{the}

\item{InnerCircle}{`list` argument which describe the circle's usage, radius, stroke,alpha, and color.}

\item{OuterCircle}{`list` argument}
}
\description{
this is in-house leaflet extension to draw cluster bubble based on bubble.
the bubble location is basically
}
